+++ src/components/header/index.tsx
@@ -6,7 +6,7 @@ import { internalLinks } from "@util/config";
 export const Header: React.FC = () => {
     return (
         <div className="header">
-            <div className="flex flex-col md:flex-row w-full md:w-1/2 max-w-4xl mx-auto">
+            <div className="flex flex-col lg:flex-row w-full lg:w-1/2 max-w-4xl mx-auto">
                 <div className="flex-grow py-2 px-10 flex flex-row justify-between items-center">
                     <CustomLink
                         href="/"
@@ -16,10 +16,10 @@ export const Header: React.FC = () => {
                     </CustomLink>
                     <DarkModeToggle />
                 </div>
-                <div className="flex-grow p-1 flex flex-row justify-evenly items-center">
+                <div className="flex-grow p-1 flex flex-row justify-center lg:justify-evenly items-center">
                     {internalLinks.map((link) => (
                         <CustomLink
-                            className="hover:text-orange dark:hover:text-purple uppercase font-grotesk text-lg font-medium"
+                            className="hover:text-orange dark:hover:text-purple mx-2 uppercase font-display text-lg font-medium"
                             key={link.text}
                             href={link.url.toString()}
                         >